Clean up ManagementClientExtensions
